Fix type doc blocks in annotation classes
