+++ lib/Doctrine/ORM/Mapping/AttributeOverride.php
@@ -17,13 +17,15 @@ final class AttributeOverride implements Annotation
      * The name of the property whose mapping is being overridden.
      *
      * @var string
+     * @readonly
      */
     public $name;
 
     /**
      * The column definition.
      *
-     * @var \Doctrine\ORM\Mapping\Column
+     * @var Column
+     * @readonly
      */
     public $column;
