+++ lib/Doctrine/ORM/Mapping/AttributeOverrides.php
@@ -6,6 +6,7 @@ namespace Doctrine\ORM\Mapping;
 
 use Attribute;
 
+use function array_values;
 use function is_array;
 
 /**
@@ -21,7 +22,8 @@ final class AttributeOverrides implements Annotation
     /**
      * One or more field or property mapping overrides.
      *
-     * @var array<AttributeOverride>
+     * @var list<AttributeOverride>
+     * @readonly
      */
     public $overrides = [];
 
@@ -36,8 +38,8 @@ final class AttributeOverrides implements Annotation
             if (! ($override instanceof AttributeOverride)) {
                 throw MappingException::invalidOverrideType('AttributeOverride', $override);
             }
-
-            $this->overrides[] = $override;
         }
+
+        $this->overrides = array_values($overrides);
     }
 }
