+++ lib/Doctrine/ORM/Mapping/Cache.php
@@ -18,14 +18,18 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 final class Cache implements Annotation
 {
     /**
+     * The concurrency strategy.
+     *
      * @Enum({"READ_ONLY", "NONSTRICT_READ_WRITE", "READ_WRITE"})
-     * @var string The concurrency strategy.
+     * @var string
+     * @psalm-var 'READ_ONLY'|'NONSTRICT_READ_WRITE'|'READ_WRITE'
      */
     public $usage = 'READ_ONLY';
 
     /** @var string|null Cache region name. */
     public $region;
 
+    /** @psalm-param 'READ_ONLY'|'NONSTRICT_READ_WRITE'|'READ_WRITE' $usage */
     public function __construct(string $usage = 'READ_ONLY', ?string $region = null)
     {
         $this->usage  = $usage;
