+++ lib/Doctrine/ORM/Mapping/ChangeTrackingPolicy.php
@@ -19,10 +19,13 @@ final class ChangeTrackingPolicy implements Annotation
      * The change tracking policy.
      *
      * @var string
+     * @psalm-var 'DEFERRED_IMPLICIT'|'DEFERRED_EXPLICIT'|'NOTIFY'
+     * @readonly
      * @Enum({"DEFERRED_IMPLICIT", "DEFERRED_EXPLICIT", "NOTIFY"})
      */
     public $value;
 
+    /** @psalm-param 'DEFERRED_IMPLICIT'|'DEFERRED_EXPLICIT'|'NOTIFY' $value */
     public function __construct(string $value)
     {
         $this->value = $value;
