+++ lib/Doctrine/ORM/Mapping/Column.php
@@ -5,29 +5,40 @@ declare(strict_types=1);
 namespace Doctrine\ORM\Mapping;
 
 use Attribute;
+use BackedEnum;
 use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 
 /**
  * @Annotation
- * @NamedArgumentConstructor()
+ * @NamedArgumentConstructor
  * @Target({"PROPERTY","ANNOTATION"})
  */
 #[Attribute(Attribute::TARGET_PROPERTY)]
 final class Column implements Annotation
 {
-    /** @var string|null */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $name;
 
-    /** @var mixed */
+    /**
+     * @var mixed
+     * @readonly
+     */
     public $type;
 
-    /** @var int|null */
+    /**
+     * @var int|null
+     * @readonly
+     */
     public $length;
 
     /**
      * The precision for a decimal (exact numeric) column (Applies only for decimal column).
      *
      * @var int|null
+     * @readonly
      */
     public $precision = 0;
 
@@ -35,40 +46,63 @@ final class Column implements Annotation
      * The scale for a decimal (exact numeric) column (Applies only for decimal column).
      *
      * @var int|null
+     * @readonly
      */
     public $scale = 0;
 
-    /** @var bool */
+    /**
+     * @var bool
+     * @readonly
+     */
     public $unique = false;
 
-    /** @var bool */
+    /**
+     * @var bool
+     * @readonly
+     */
     public $nullable = false;
 
-    /** @var bool */
+    /**
+     * @var bool
+     * @readonly
+     */
     public $insertable = true;
 
-    /** @var bool */
+    /**
+     * @var bool
+     * @readonly
+     */
     public $updatable = true;
 
-    /** @var class-string<\BackedEnum>|null */
+    /**
+     * @var class-string<BackedEnum>|null
+     * @readonly
+     */
     public $enumType = null;
 
-    /** @var array<string,mixed> */
+    /**
+     * @var array<string,mixed>
+     * @readonly
+     */
     public $options = [];
 
-    /** @var string|null */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $columnDefinition;
 
     /**
      * @var string|null
+     * @readonly
      * @psalm-var 'NEVER'|'INSERT'|'ALWAYS'|null
      * @Enum({"NEVER", "INSERT", "ALWAYS"})
      */
     public $generated;
 
     /**
-     * @param class-string<\BackedEnum>|null $enumType
-     * @param array<string,mixed>            $options
+     * @param class-string<BackedEnum>|null $enumType
+     * @param array<string,mixed>           $options
      * @psalm-param 'NEVER'|'INSERT'|'ALWAYS'|null $generated
      */
     public function __construct(
