+++ lib/Doctrine/ORM/Mapping/DiscriminatorMap.php
@@ -15,7 +15,10 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 #[Attribute(Attribute::TARGET_CLASS)]
 final class DiscriminatorMap implements Annotation
 {
-    /** @var array<int|string, string> */
+    /**
+     * @var array<int|string, string>
+     * @readonly
+     */
     public $value;
 
     /** @param array<int|string, string> $value */
