+++ lib/Doctrine/ORM/Mapping/Embedded.php
@@ -15,10 +15,16 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 #[Attribute(Attribute::TARGET_PROPERTY)]
 final class Embedded implements Annotation
 {
-    /** @var string|null */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $class;
 
-    /** @var string|bool|null */
+    /**
+     * @var string|bool|null
+     * @readonly
+     */
     public $columnPrefix;
 
     public function __construct(?string $class = null, $columnPrefix = null)
