+++ lib/Doctrine/ORM/Mapping/GeneratedValue.php
@@ -16,13 +16,16 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 final class GeneratedValue implements Annotation
 {
     /**
-     * The type of Id generator.
+     * The type of ID generator.
      *
      * @var string
+     * @psalm-var 'AUTO'|'SEQUENCE'|'IDENTITY'|'NONE'|'CUSTOM'
+     * @readonly
      * @Enum({"AUTO", "SEQUENCE", "TABLE", "IDENTITY", "NONE", "UUID", "CUSTOM"})
      */
     public $strategy = 'AUTO';
 
+    /** @psalm-param 'AUTO'|'SEQUENCE'|'IDENTITY'|'NONE'|'CUSTOM' $strategy */
     public function __construct(string $strategy = 'AUTO')
     {
         $this->strategy = $strategy;
