+++ lib/Doctrine/ORM/Mapping/InheritanceType.php
@@ -19,10 +19,13 @@ final class InheritanceType implements Annotation
      * The inheritance type used by the class and its subclasses.
      *
      * @var string
+     * @psalm-var 'NONE'|'JOINED'|'SINGLE_TABLE'|'TABLE_PER_CLASS'
+     * @readonly
      * @Enum({"NONE", "JOINED", "SINGLE_TABLE", "TABLE_PER_CLASS"})
      */
     public $value;
 
+    /** @psalm-param 'NONE'|'JOINED'|'SINGLE_TABLE'|'TABLE_PER_CLASS' $value */
     public function __construct(string $value)
     {
         $this->value = $value;
