+++ lib/Doctrine/ORM/Mapping/JoinTable.php
@@ -15,19 +15,34 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 #[Attribute(Attribute::TARGET_PROPERTY)]
 final class JoinTable implements Annotation
 {
-    /** @var string|null */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $name;
 
-    /** @var string|null */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $schema;
 
-    /** @var array<\Doctrine\ORM\Mapping\JoinColumn> */
+    /**
+     * @var array<JoinColumn>
+     * @readonly
+     */
     public $joinColumns = [];
 
-    /** @var array<\Doctrine\ORM\Mapping\JoinColumn> */
+    /**
+     * @var array<JoinColumn>
+     * @readonly
+     */
     public $inverseJoinColumns = [];
 
-    /** @var array<string, mixed> */
+    /**
+     * @var array<string, mixed>
+     * @readonly
+     */
     public $options = [];
 
     /** @param array<string, mixed> $options */
