+++ lib/Doctrine/ORM/Mapping/ManyToOne.php
@@ -15,26 +15,38 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 #[Attribute(Attribute::TARGET_PROPERTY)]
 final class ManyToOne implements Annotation
 {
-    /** @var class-string|null */
+    /**
+     * @var class-string|null
+     * @readonly
+     */
     public $targetEntity;
 
-    /** @var string[]|null */
+    /**
+     * @var string[]|null
+     * @readonly
+     */
     public $cascade;
 
     /**
      * The fetching strategy to use for the association.
      *
      * @var string
+     * @psalm-var 'LAZY'|'EAGER'|'EXTRA_LAZY'
+     * @readonly
      * @Enum({"LAZY", "EAGER", "EXTRA_LAZY"})
      */
     public $fetch = 'LAZY';
 
-    /** @var string|null */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $inversedBy;
 
     /**
      * @param class-string|null $targetEntity
      * @param string[]|null     $cascade
+     * @psalm-param 'LAZY'|'EAGER'|'EXTRA_LAZY' $fetch
      */
     public function __construct(
         ?string $targetEntity = null,
