+++ lib/Doctrine/ORM/Mapping/OneToMany.php
@@ -15,32 +15,50 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 #[Attribute(Attribute::TARGET_PROPERTY)]
 final class OneToMany implements Annotation
 {
-    /** @var string */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $mappedBy;
 
-    /** @var class-string|null */
+    /**
+     * @var class-string|null
+     * @readonly
+     */
     public $targetEntity;
 
-    /** @var array<string> */
+    /**
+     * @var array<string>|null
+     * @readonly
+     */
     public $cascade;
 
     /**
      * The fetching strategy to use for the association.
      *
      * @var string
+     * @psalm-var 'LAZY'|'EAGER'|'EXTRA_LAZY'
+     * @readonly
      * @Enum({"LAZY", "EAGER", "EXTRA_LAZY"})
      */
     public $fetch = 'LAZY';
 
-    /** @var bool */
+    /**
+     * @var bool
+     * @readonly
+     */
     public $orphanRemoval = false;
 
-    /** @var string */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $indexBy;
 
     /**
      * @param class-string|null $targetEntity
      * @param string[]|null     $cascade
+     * @psalm-param 'LAZY'|'EAGER'|'EXTRA_LAZY' $fetch
      */
     public function __construct(
         ?string $mappedBy = null,
