+++ lib/Doctrine/ORM/Mapping/SequenceGenerator.php
@@ -15,13 +15,22 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 #[Attribute(Attribute::TARGET_PROPERTY)]
 final class SequenceGenerator implements Annotation
 {
-    /** @var string */
+    /**
+     * @var string|null
+     * @readonly
+     */
     public $sequenceName;
 
-    /** @var int */
+    /**
+     * @var int
+     * @readonly
+     */
     public $allocationSize = 1;
 
-    /** @var int */
+    /**
+     * @var int
+     * @readonly
+     */
     public $initialValue = 1;
 
     public function __construct(
